scheduler: generalization logic from le_renewer class
